1906 - change check owner

+++ src/Application/Command/Sheet/LinkedSheets/DeleteHandler.php
@@ -2,6 +2,8 @@
 
 namespace Proximum\Vimeet\Application\Command\Sheet\LinkedSheets;
 
+use Proximum\Vimeet\Application\Criteria\LinkedSheets\AreRemovableLinkedSheetsCriteria;
+use Proximum\Vimeet\Domain\Exception\DomainException;
 use Proximum\Vimeet\Domain\Repository\Sheet\LinkedSheetsRepositoryInterface;
 
 class DeleteHandler
@@ -9,13 +11,23 @@ class DeleteHandler
     /** @var LinkedSheetsRepositoryInterface */
     private $linkedSheetsRepository;
 
-    public function __construct(LinkedSheetsRepositoryInterface $linkedSheetsRepository)
-    {
+    /** @var AreRemovableLinkedSheetsCriteria */
+    private $areRemovableLinkedSheetsCriteria;
+
+    public function __construct(
+        LinkedSheetsRepositoryInterface $linkedSheetsRepository,
+        AreRemovableLinkedSheetsCriteria $areRemovableLinkedSheetsCriteria
+    ) {
         $this->linkedSheetsRepository = $linkedSheetsRepository;
+        $this->areRemovableLinkedSheetsCriteria = $areRemovableLinkedSheetsCriteria;
     }
 
     public function handle(Delete $command)
     {
+        if (count($this->areRemovableLinkedSheetsCriteria->meetCriteria([$command->linkedSheets])) === 0) {
+            throw new DomainException('Linked sheets can\'t be removed');
+        }
+
         $this->linkedSheetsRepository->remove($command->linkedSheets);
     }
 }
